+++ ChangeLog.md
@@ -71,6 +71,7 @@
 * Improve handling of WSI creation: faster offset table calucation and a naming of temp files to allow more than 64.000.
 * Change: DicomAnonymizer private fields and methods changed to protected so they can be used in subclasses, made instance methods virtual so they can be overridden in subclasses
 * Fix VR's SV and UV VR Length field (#1386)
+* Add support for rendering multiframe DICOM files where last fragment is 0 bytes (#1586) 
 
 #### 5.0.2 (2022-01-11)
 * Update to DICOM Standard 2021e